レビュー後修正

+++ app/controllers/items_controller.rb
@@ -38,7 +38,9 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    redirect_to root_path if @item.destroy
+    unless @item.user_id == current_user.id
+      redirect_to root_path if @item.destroy
+    end
   end
 
 end
